added a device as a new pag
