+++ App.js
@@ -22,6 +22,7 @@ import { LogBox } from 'react-native'
 import Swiper from './screen/Swiper'
 import Monitor from './screen/Monitor'
 import Room from './screen/Room'
+import Device from './screen/Device'
 LogBox.ignoreLogs(['Warning: ...'])
 LogBox.ignoreAllLogs()
 
@@ -139,6 +140,7 @@ const App = () => {
           <Stack.Screen name="Splash" component={Splash} />
           <Stack.Screen name="Swiper" component={Swiper} />
           <Stack.Screen name="Room" component={Room} />
+          <Stack.Screen name="Device" component={Device} />
         </Stack.Navigator>
       </KeyboardAvoidingView>
     </NavigationContainer>
